Fix macOS build: mode_t variadic promotion, cfg-gated unused vars
- Cast mode_t to c_uint for openat() variadic arg (mode_t is u16 on
  macOS, variadic C functions require at least c_uint)
- Prefix timeout and network_pool with underscore where they're only
  used inside #[cfg(target_os = "linux")] blocks

+++ src/io/threadpool.rs
@@ -374,7 +374,12 @@ impl ThreadPoolBackend {
                 }
                 PoolOp::Open { path, flags, mode } => {
                     let fd = unsafe {
-                        libc::openat(libc::AT_FDCWD, path.as_ptr(), flags, mode as libc::mode_t)
+                        libc::openat(
+                            libc::AT_FDCWD,
+                            path.as_ptr(),
+                            flags,
+                            libc::c_uint::from(mode as libc::mode_t),
+                        )
                     };
                     if fd < 0 {
                         (
